Convert AppIntroBase to Kotlin
This is a major refactoring of the AppIntroBase class. I've converted
the class to Kotlin and I've cleaned up a lot of deprecated/unused/dead
code.

I've reduced the surface of the public API to expose only what is really
necessary (the `protected` methods) and make private all the
implementaiton details.

Summary of changes:
* Refactored Callback (removed deprecated callbacks).
* Reduced number of boolean flags (now every flag is a kotlin property
so is easier to understand the side effects).
* Introduced the `onIntroFinished` and `onNextSlide` callback, this
simplifies a bit all the callback code.
* Refactored all the Immersive Mode code to only enable Sticky Immersive
as the non-Sticky made no sense for this library.
* Cleaned up the `checkAndRequestPermissions` to use a more concise functional
approach.