Fix bug on swipe with Permission slide
Currently the Swiping on Permission slide is bugged.
When swiping backwards and with a permission slides, it could happen
that the carousel is left "frozen" in between two slides. I'm fixing it.

Moreover I'm also unlocking the slide when the user denied a permission.
It should still be possible for a user to eventually go back in
the carousel even if they rejected a permission.